Move config into own module
